Added 1. putN to put numbres with recursion as char 2. print_unsigned integers
