+++ holberton.h
@@ -24,6 +24,7 @@ typedef struct Ftype f_type;
 int _printf(const char *format, ...);
 int operator_printf(char s_type, va_list list);
 int _putchar(char c);
+int putN(unsigned int n, unsigned int b, char *nums)
 
 /* functions from standar library in f_library like putchar and others*/
